+++ src/Factories/RequestFactoryV24.php
@@ -158,6 +158,16 @@ final class RequestFactoryV24 extends RequestFactoryV2
         throw new LogicException('Method not implemented yet for EBICS 2.4');
     }
 
+    public function createXEK(
+        DateTimeInterface $dateTime,
+        DateTimeInterface $startDateTime = null,
+        DateTimeInterface $endDateTime = null,
+        int $segmentNumber = null,
+        bool $isLastSegment = null
+    ): Request {
+        throw new LogicException('Method not implemented yet for EBICS 2.4');
+    }
+
     public function createCCT(DateTimeInterface $dateTime, UploadTransaction $transaction, bool $withES): Request
     {
         throw new LogicException('Method not implemented yet for EBICS 2.4');